requirements.txt
Pillow==10.3.0
exifread==3.0.0
imagehash==4.3.1